+++ Classes/ItemStorage.lua
@@ -37,8 +37,18 @@ local error, table, tostring, tinsert, tremove, type, select, FindInTableIf, tim
 
 function addon:InitItemStorage()-- Extract items from our SV. Could be more elegant
    db = self:Getdb()
+   local stored;
    for k, v in ipairs(db.itemStorage) do
-      Storage:StoreItem(v.link, v.type, "restored", v)
+      stored = Storage:StoreItem(v.link, v.type, "restored", v)
+      if not stored then -- Item probably no longer exists?
+         addon:Debug("Error - ItemsStorage, couldn't add db item:", v.link)
+         local key = FindInTableIf(db.itemStorage, function(d) return addon:ItemIsItem(v.link, d.link) end)
+         if key then
+            tremove(db.itemStorage, key)
+         else
+            addon:Debug("Error - Unable to remove item from db?!")
+         end
+      end
    end
 end
 
@@ -58,7 +68,12 @@ local function findItemInBags(link)
 end
 
 local function newItem(link, type, time_remaining)
-   local Item = setmetatable({}, item_prototype)
+   local Item = setmetatable({}, {
+      __index = item_prototype,
+      __tostring = function(self)
+         return self.link
+      end,
+   })
    Item.link = link
    Item.type = type and type or Item.type
    Item.time_remaining = time_remaining and time_remaining or Item.time_remaining